v3 Phase 1: dict cap 65K → 262K, u16 → u32 IDs (Δ -0.031 bpb on full panel).
Following the v3 survey-first methodology, the first architectural
change informed by measurement: widen DICT_CAPACITY from 1<<16 to
1<<18, switch dict-id type from u16 to u32 throughout, emit 18 bits
per id, bump ID_BIT_K from 22 to 24 (4M → 16M context slots) to
keep collision rate near zero at the new context-space size.

Pre-measurement prediction (from survey, biased toward floor):
  0.3-0.5 bpb expected gain.
Actual on full 20-window enwik8 panel:
  2.381 bpb, Δ -0.031 vs Phase 16's 2.412.

Real but ~10x below predicted — the survey methodology was wrong
in a specific way worth journaling. The survey measured OOV rate
on a fresh walk from byte 0; the bench panel uses 4 MiB warm
prefix that already absorbs most "first-N-occurrence" tokens
before measure starts. The panel's effective OOV rate is
dominated by first-occurrence-in-corpus proper nouns specific to
the article in this window — those are OOV at any dict cap.

Decomposition shift (quick-panel, per 64 KiB window avg):
  token_oov_ac:   28,600 → 27,460 bits  (-1,140)
  token_hit_ac:   45,000 → 45,200       (+200)
  lz_match_ac:    80,000 → 80,200       (+200)
  others:         barely changed

Net ~0.04 bpb saved in OOV bits, partially offset by tiny
overheads elsewhere — consistent with the 0.031 bpb headline.

Implementation: mechanical u16→u32 refactor across xml_tok.rs and
tok_lz.rs. Length encoding stays 16-bit. All 7 xml-tok roundtrip
tests pass incl. real enwik8; 129 tests total.

Calibration update for future predictions: when the survey suggests
N bpb headroom from a single targeted change, divide by 5-10x to
get the realistic codec-level gain. The survey's value is direction
(word > BPE, conditional > Order-0); magnitude needs further
discount.

Next: Order-2 word context. Literature: Order-1 → Order-2 cuts
conditional entropy 20-30%. Calibration-corrected expectation:
0.05-0.15 bpb gain, landing ~2.23-2.33 bpb.

Also queued: re-run the survey under panel-style prewarm conditions
so predictions actually match what the codec sees.

JOURNAL.md updated with the calibration analysis and the third
survey blind spot (static-vs-adaptive-with-prewarm mismatch).

+++ src/tok_lz.rs
@@ -42,7 +42,7 @@ pub(crate) struct TokenMatcher {
     /// All token ids emitted to the Content stream so far, in
     /// emission order. Both warm and measure tokens land here; the
     /// matcher walks back up to `WINDOW_SIZE` positions.
-    stream: Vec<u16>,
+    stream: Vec<u32>,
 }
 
 impl TokenMatcher {
@@ -60,14 +60,14 @@ impl TokenMatcher {
     }
 
     /// Token id at stream position `pos`.
-    pub(crate) fn at(&self, pos: usize) -> u16 {
+    pub(crate) fn at(&self, pos: usize) -> u32 {
         self.stream[pos]
     }
 
     /// Append a token id to the stream. After the third token,
     /// every appended token closes a `MIN_MATCH`-sized window whose
     /// hash is inserted at the head of its bucket's chain.
-    pub(crate) fn push(&mut self, id: u16) {
+    pub(crate) fn push(&mut self, id: u32) {
         self.stream.push(id);
         let n = self.stream.len();
         if n >= MIN_MATCH {
@@ -88,7 +88,7 @@ impl TokenMatcher {
         h
     }
 
-    fn hash_slice(slice: &[u16]) -> u64 {
+    fn hash_slice(slice: &[u32]) -> u64 {
         debug_assert!(slice.len() >= MIN_MATCH);
         let mut h = FNV_OFFSET;
         for &t in &slice[..MIN_MATCH] {
@@ -101,7 +101,7 @@ impl TokenMatcher {
     /// the stream's history. Returns `(offset_in_tokens, length)`
     /// or `None` when no `MIN_MATCH`-clearing match exists within
     /// the `WINDOW_SIZE` lookback.
-    pub(crate) fn find_match(&self, upcoming: &[u16]) -> Option<(u32, u32)> {
+    pub(crate) fn find_match(&self, upcoming: &[u32]) -> Option<(u32, u32)> {
         if upcoming.len() < MIN_MATCH {
             return None;
         }
@@ -165,7 +165,7 @@ mod tests {
     #[test]
     fn finds_simple_repeat() {
         let mut m = TokenMatcher::new();
-        for id in [1u16, 2, 3, 4, 5] {
+        for id in [1u32, 2, 3, 4, 5] {
             m.push(id);
         }
         // Looking for "1 2 3 4 ..." should find offset 5 (back to the start).
@@ -178,7 +178,7 @@ mod tests {
     #[test]
     fn no_match_below_min() {
         let mut m = TokenMatcher::new();
-        for id in [1u16, 2, 3, 4, 5] {
+        for id in [1u32, 2, 3, 4, 5] {
             m.push(id);
         }
         // "1 2 9 ..." has only 2 matching tokens, below MIN_MATCH=3.
@@ -189,13 +189,13 @@ mod tests {
     fn longer_match_preferred() {
         let mut m = TokenMatcher::new();
         // Two earlier matches of "1 2 3": one extends 3 tokens, one extends 5.
-        for id in [1u16, 2, 3, 99, 99] {
+        for id in [1u32, 2, 3, 99, 99] {
             m.push(id);
         }
-        for id in [1u16, 2, 3, 4, 5] {
+        for id in [1u32, 2, 3, 4, 5] {
             m.push(id);
         }
-        for id in [77u16, 77, 77] {
+        for id in [77u32, 77, 77] {
             m.push(id);
         }
         // Now look for "1 2 3 4 5 ..." — the second occurrence is 5 long.
